changed preprocess_data.py into build_features.py and created model_definition script
