+++ config.yml
@@ -11,3 +11,14 @@ processed_paths:
 
 params:
   sequence_length: 200
+  loss_function: "binary_crossentropy"
+  optimizer: "adam"
+  batch_train: 5000
+  batch_test: 5000
+  categories:
+    - "phishing"
+    - "legitimate"
+  char_index: null  # This will be updated after initializing the tokenizer
+  epoch: 30
+  embedding_dimension: 50
+  dataset_dir: "../dataset/small_dataset/"
