+++ src/models/model_definition.py
@@ -1,0 +1,40 @@
+# src/models/model_definition.py
+
+from keras.models import Sequential
+from keras.layers import Embedding, Conv1D, MaxPooling1D, Flatten, Dense, Dropout
+
+
+def build_model(voc_size, embedding_dim, categories):
+    model = Sequential()
+    voc_size = len(char_index.keys())
+    model.add(Embedding(voc_size + 1, 50))
+
+    model.add(Conv1D(128, 3, activation='tanh'))
+    model.add(MaxPooling1D(3))
+    model.add(Dropout(0.2))
+
+    model.add(Conv1D(128, 7, activation='tanh', padding='same'))
+    model.add(Dropout(0.2))
+
+    model.add(Conv1D(128, 5, activation='tanh', padding='same'))
+    model.add(Dropout(0.2))
+
+    model.add(Conv1D(128, 3, activation='tanh', padding='same'))
+    model.add(MaxPooling1D(3))
+    model.add(Dropout(0.2))
+
+    model.add(Conv1D(128, 5, activation='tanh', padding='same'))
+    model.add(Dropout(0.2))
+
+    model.add(Conv1D(128, 3, activation='tanh', padding='same'))
+    model.add(MaxPooling1D(3))
+    model.add(Dropout(0.2))
+
+    model.add(Conv1D(128, 3, activation='tanh', padding='same'))
+    model.add(MaxPooling1D(3))
+    model.add(Dropout(0.2))
+
+    model.add(Flatten())
+    model.add(Dense(len(params['categories'])-1, activation='sigmoid'))
+
+    model.compile(loss=)
